# Simplify HTML escaping