close https://github.com/haml/haml/issues/901

All details are described in https://github.com/haml/haml/issues/901.

## Benchmark
Without implementing Temple compiler for attributes, it improves performance.

With Ruby 2.4.0 and  [k0kubun/haml_bench/templates/slim_bench.haml](https://github.com/k0kubun/haml_bench/blob/97a7b9970684b975c48e1403ee4fe329bf1151fb/templates/slim_bench.haml),

### before
```
$ bundle exec ruby bench.rb
Rendering: /home/k0kubun/src/github.com/k0kubun/haml_bench/templates/slim_bench.haml
Calculating -------------------------------------
          haml 4.0.7     4.183k i/100ms
   haml 5.0.0.beta.2     5.385k i/100ms
-------------------------------------------------
          haml 4.0.7     43.307k (± 1.0%) i/s -    217.516k
   haml 5.0.0.beta.2     56.453k (± 1.5%) i/s -    285.405k

Comparison:
   haml 5.0.0.beta.2:    56453.0 i/s
          haml 4.0.7:    43307.1 i/s - 1.30x slower
```

### after
```
$ bundle exec ruby bench.rb
Rendering: /home/k0kubun/src/github.com/k0kubun/haml_bench/templates/slim_bench.haml
Calculating -------------------------------------
          haml 4.0.7     4.178k i/100ms
   haml 5.0.0.beta.2     5.827k i/100ms
-------------------------------------------------
          haml 4.0.7     43.184k (± 1.3%) i/s -    217.256k
   haml 5.0.0.beta.2     61.642k (± 1.5%) i/s -    308.831k

Comparison:
   haml 5.0.0.beta.2:    61642.4 i/s
          haml 4.0.7:    43184.0 i/s - 1.43x slower
```