@k0kubun 2017-02-08
Since @teeparham reacted :+1:, and to move it forward, merging this change for now. Feel free to revert this branch if you don't want to add this change in Haml 5.0.0.